Version 1.7.0 of the AWS Java SDK - This release adds support for Amazon SQS Dead Letter Queues, new Amazon EC2 instance types and changes the way AWS region metadata is looked up.
See the full release notes at http://aws.amazon.com/releasenotes/4084807713219428

+++ src/main/java/com/amazonaws/util/ContentLengthValidationInputStream.java
@@ -31,13 +31,13 @@ public class ContentLengthValidationInputStream extends FilterInputStream {
 	private long expectedLength;
 
 	/** This variables holds the number of bytes read from the stream so far. */
-	private int bytesRead;
+	private long bytesRead;
 
 	/**
 	 * This variable holds the number of bytes read from the stream at the
 	 * instance when mark method was called.
 	 */
-	private int mark;
+	private long mark;
 
 	public ContentLengthValidationInputStream(InputStream in, long expectedLength) {
 		super(in);
